// src/app/core/interfaces/common.enums.ts
export enum Gender {
  Male = 'male',
  Female = 'female',
}

export enum FormMode {
  Add = 'add',
  Edit = 'edit',
}

export enum WidgetType {
  Advertisement = 'advertisement',
  Newsletter = 'newsletter',
  TicketManagement = 'ticket-management',
}

export enum TicketStatus {
  Pending = 1,
  Open,
  Close,
}